// clients/Client3/assets/src/modules/Battle/directions/Direction8.ts
/**
 * 8方向枚举
 */
export enum Direction8 {
    /**
     * 南(下)
     */
    B_B = 1,
    /**
     * 东南(右下)
     */
    R_B = 2,
    /**
     * 东(右)
     */
    R_R = 3,
    /**
     * 东北(右上)
     */
    R_T = 4,
    /**
     * 北(上)
     */
    T_T = 5,
    /**
     * 西北(左上)
     */
    L_T = 6,
    /**
     * 西(左)
     */
    L_L = 7,
    /**
     * 西南(左下)
     */
    L_B = 8,
}